+++ src/steps/base/simple/external_workflow_executor.ts
@@ -2,10 +2,10 @@ import { Logger } from 'pino';
 import { join } from 'path';
 import { WorkflowUtils } from '../../../utils';
 import { WorkflowEngine } from '../../../workflow';
-import { Dictionary, ExecutionBindings, Workflow } from '../../../types';
+import { Dictionary, ExecutionBindings } from '../../../types';
 import { BaseStepExecutor } from '../base_executor';
 import { SimpleStep, StepOutput } from '../../types';
-import { WorkflowEngineError } from '../../../errors';
+import { StepCreationError } from '../../../errors';
 
 export class ExternalWorkflowStepExecutor extends BaseStepExecutor {
   private readonly workflowEngine: WorkflowEngine;
@@ -17,14 +17,10 @@ export class ExternalWorkflowStepExecutor extends BaseStepExecutor {
 
   private prepareExternalWorkflowEngine(step: SimpleStep): WorkflowEngine {
     if (!step.externalWorkflow) {
-      throw new WorkflowEngineError(
-        'externalWorkflow required for ExternalWorkflowStepExecutor',
-        400,
-        step.name,
-      );
+      throw new StepCreationError('externalWorkflow required', step.name);
     }
     const workflowPath = join(this.rootPath, step.externalWorkflow.path);
-    const workflow = WorkflowUtils.createFromFilePath<Workflow>(workflowPath);
+    const workflow = WorkflowUtils.createWorkflowFromFilePath(workflowPath);
     const externalWorkflowRootPath = join(this.rootPath, step.externalWorkflow.rootPath || '');
     return new WorkflowEngine(workflow, externalWorkflowRootPath);
   }