+++ src/steps/base/simple/factory.ts
@@ -1,10 +1,10 @@
 import { Logger } from 'pino';
-import { WorkflowEngineError } from '../../../errors';
+import { StepCreationError, WorkflowCreationError } from '../../../errors';
 import { Dictionary } from '../../../types';
 import { SimpleStep, StepExecutor } from '../../types';
 import { BaseStepExecutor } from '../base_executor';
 import { ExternalWorkflowStepExecutor } from './external_workflow_executor';
-import { FunctionStepExecutor } from './function_exector';
+import { FunctionStepExecutor } from './function_executor';
 import { TemplateStepExecutor } from './template_executor';
 
 export class SimpleStepExecutorFactory {
@@ -27,6 +27,6 @@ export class SimpleStepExecutorFactory {
       return new TemplateStepExecutor(step, rootPath, bindings, simpleStepLogger);
     }
 
-    throw new WorkflowEngineError('Invalid simple step configuration', 400, step.name);
+    throw new StepCreationError('Invalid simple step configuration', step.name);
   }
 }